1.0

+++ src/mfm/tools/Tools.java
@@ -1,6 +1,7 @@
 package mfm.tools;
 
 import java.io.File;
+import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -9,11 +10,21 @@ import java.util.Comparator;
 import java.util.Scanner;
 
 public class Tools {
+	
+	//add printwriter
+	//add buffreader
 
 	public static String scan() {
 		Scanner sc = new Scanner(System.in);
-		String out = sc.nextLine();
-		return out;
+		return sc.nextLine();
+	}
+	
+	//7z
+	
+	public static void write7z(String name, String txt) throws IOException{
+		FileWriter write7z = new FileWriter(name);
+		write7z.write(txt);
+		write7z.close();
 	}
 	
 	public static void run7z(String in) throws InterruptedException, IOException {
@@ -24,24 +35,31 @@ public class Tools {
 			e.printStackTrace();
 			System.out.println("you need to download 7-Zip!  7-zip.org");
 		}
-		deleteAll("temp");
+		deleteAll("temp",null);
 	}
+	
+	//substring
+	
 	public static String removeLast(String var, int remove) {
 		var = var.substring(0,var.length()-remove);
 		return var;
 	}
 	
-	public static void deleteAll(String loc) throws IOException {
+	//delete
+	
+	public static void deleteAll(String loc, String notThisOne) throws IOException {
 		Path dir = Paths.get(loc);
 		if (dir.toFile().exists()) {
 			Files.walk(dir)
 					.sorted(Comparator.reverseOrder())
 					.forEach(path -> {
-						try {
-							System.out.println("Deleting: " + path);
-							Files.delete(path);
-						} catch (IOException e) {
-							e.printStackTrace();
+						if (notThisOne != path.toString()) {	//check if path is complet so it works
+							try {
+								System.out.println("Deleting: " + path);
+								Files.delete(path);
+							} catch (IOException e) {
+								e.printStackTrace();
+							}
 						}
 					});
 		}
@@ -49,6 +67,8 @@ public class Tools {
 	public static void deleteFile(String loc) throws IOException {
 		new File(loc).delete();
 	}
+	
+	//available
 	
 	public static String[] available (String in, boolean all, boolean print) {
 		if (print) {
